Create game class, author class, and their methods
